requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pypi_0
astetik=1.13=pypi_0
astor=0.8.1=pypi_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pypi_0
boto3=1.23.10=pypi_0
botocore=1.26.10=pypi_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.10.26=h06a4308_2
cached-property=1.5.2=pypi_0
cachetools=4.2.4=pypi_0
certifi=2020.6.20=pyhd3eb1b0_3
chances=0.1.9=pypi_0
charset-normalizer=2.0.12=pypi_0
clang=5.0=pypi_0
click=8.0.4=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.30=pypi_0
dataclasses=0.8=pypi_0
decorator=4.4.2=pypi_0
efficientnet=1.0.0=pypi_0
et-xmlfile=1.1.0=pypi_0
filelock=3.4.1=pypi_0
fitter=1.5.2=pypi_0
flatbuffers=1.12=pypi_0
gast=0.4.0=pypi_0
gensim=3.8.3=pypi_0
geonamescache=1.3.0=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.42.0=pypi_0
h5py=3.1.0=pypi_0
hdf5=1.12.1=h69dfa17_1
idna=3.3=pypi_0
imageio=2.13.2=pypi_0
importlib-metadata=4.8.2=pypi_0
importlib-resources=5.4.0=pypi_0
ipython=7.16.3=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.2=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.1.0=pypi_0
keras=2.6.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kerasplotlib=1.0=pypi_0
kiwisolver=1.3.1=pypi_0
krb5=1.19.2=hac12032_0
ld_impl_linux-64=2.35.1=h7274673_9
libcurl=7.78.0=h0b77cf5_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libnghttp2=1.46.0=hce63b2e_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.3.0=hd4cf53a_17
markdown=3.3.6=pypi_0
matplotlib=3.3.4=pypi_0
ncurses=6.3=h7f8727e_2
networkx=2.5.1=pypi_0
nltk=3.4.5=pypi_0
numpy=1.19.5=pypi_0
oauthlib=3.2.0=pypi_0
opencv-python=4.5.4.60=pypi_0
openpyxl=3.1.2=pypi_0
openslide-python=1.1.2=pypi_0
openssl=1.1.1l=h7f8727e_0
opt-einsum=3.3.0=pypi_0
pandas=1.1.5=pypi_0
parso=0.7.1=pypi_0
patsy=0.5.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=pypi_0
pip=21.2.2=py36h06a4308_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.19.1=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydicom=2.2.2=pypi_0
pygments=2.12.0=pypi_0
pyparsing=3.0.6=pypi_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pypi_0
python-version=0.0.2=pypi_0
pytz=2022.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.1=h27cfd23_0
regex=2022.6.2=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
s3transfer=0.5.2=pypi_0
sacremoses=0.0.53=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=58.0.4=py36h06a4308_0
six=1.15.0=pypi_0
sklearn=0.0=pypi_0
smart-open=6.0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
statsmodels=0.12.2=pypi_0
talos=1.3=pypi_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.6.2=pypi_0
tensorflow-estimator=2.6.0=pypi_0
termcolor=1.1.0=pypi_0
theano=1.0.5=pypi_0
threadpoolctl=3.0.0=pypi_0
tifffile=2020.9.3=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.5.0=pypi_0
torch=1.10.0=pypi_0
tqdm=4.64.0=pypi_0
traitlets=4.3.3=pypi_0
transformers=2.5.0=pypi_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.9=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wrangle=0.7.2=pypi_0
wrapt=1.12.1=pypi_0
xdg=5.1.1=pypi_0
xlrd=1.2.0=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
